requirements.txt
fastapi
uvicorn
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0
fastmcp>=3.0.0
langgraph
langchain-openai
pydantic-settings
qdrant-client
httpx
python-dotenv
python-multipart
openpyxl
pandas
joserfc